# bankAccount-domain/src/main/java/fr/kata/bankAccount/domain/data/Balance.java
package fr.kata.bankAccount.domain.data;

import lombok.Builder;
import lombok.Data;

import java.math.BigDecimal;

@Builder
@Data
public class Balance {
    private BigDecimal totalBalance;
    private String currency;
}